refactor(Constraint): Create constraint authoring component
#1127
